Move DataHubSystemAdministrator to Core
